1.2.4: Fix KeyNotFoundException in ArrowCounter
M&B code surprises once again. When agents are built, we store
a key-value for their hashcode to remaining arrows. When arrows are
fired, we look up the key for this agent and subtract from their
current arrow count.
Apparently, this sometimes results in a KeyNotFoundException.

The only explanation I can think of is:
1. HashCodes of agents are changing (big no-no!)
2. MissionBehavior.OnAgentBuilt does not get called for every
   agent that is built.

Both alternatives are quite sad. To fix, I've added a simple
ContainsKey check before querying the map. If the key is not
present, we simply do not do the map-dependent operation. This may
mean some inaccurate arrow counting when it comes to agents picking
up arrows. However, I myself have played the game for hours without
seeing this bug, so I think it is quite rare and any adverse
effects of fixing it like this will be minimal. If I get more time,
I may investigate more in-depth.