1.3.0: Make compatible with game version 1.5.9.
With this new game version, entering a battle would throw a "Method not
found" exception for GauntletLayer::LoadMovie(String, ViewModel).

Not 100% why - this method is still present in 1.5.9 and after a rebuild
with the newest DLLs, the mod is once again working correctly. Seems a
rebuild is all that was necessary.

This commit thus just bumps the version and cleans up a couple of minor
things in Config.cs.

Note regarding 1.3.0 instead of 1.2.10:
As part of this release, I intend to open source this mod. This seems
like a big enough event for us to mark it with a greater leap in version
numbering (which, to be honest, is relatively arbitrary anyway!).