1.6.1: Make compatible with Steam Workshop
Well, technically this release is not.
Steam Workshop has been released but the way that Steam installs mods
means that the DLL is not placed within the game base folder, and so the
method we currently have of finding the config.xml file does not work
and we get a file not found exception.

The (hacky) solution so I can quickly get the Steam Workshop version
working is to not allow customization and don't even try to load any
config file - just use default values. This means I have two versions
released - a Nexus version which does the config reading, and a Steam
version which does not. This commit is the Nexus version - the Steam
version simply has the bool inside Config.cs flipped to true.

I'm doing this hacky solution cause I'm low on time but next release,
let's see if we can't find another way to find the config file and thus
avoid this split-release and allow customization via the Steam version
too.